Add scoreboard countdown and login improvements
- Add retro scoreboard countdown component
- Add seed script for testing leaderboard
- Update LoginSection styling
