+++ src/components/scoreboard/ScoreboardSeparator.tsx
@@ -1,0 +1,8 @@
+export function ScoreboardSeparator() {
+	return (
+		<div className="scoreboard-separator">
+			<span className="separator-dot" />
+			<span className="separator-dot" />
+		</div>
+	);
+}
